fix remove cross after remove self from exfee, refresh cross after load user with identitis change

+++ EXFE/EFAPIServer+Crosses.h
@@ -11,7 +11,7 @@
 @class Cross;
 @interface EFAPIServer (Crosses)
 
-- (void)loadCrossWithCrossId:(int)corss_id
+- (void)loadCrossWithCrossId:(NSUInteger)corss_id
                  updatedtime:(NSDate*)updatedtime
                      success:(void (^)(RKObjectRequestOperation *operation, RKMappingResult *mappingResult))success
                      failure:(void (^)(RKObjectRequestOperation *operation, NSError *error))failure;
